ecosip-ai-backend/eco-impact-tracker/requirements.txt
streamlit
opencv-python
numpy
google-generativeai
python-dotenv
Pillow